The rebuilding of Aza
After the failed attempt of trying to convert to unity 2019.3, alot of rebuilding had to take place. After that, jumping and air combat was began and jumping is now perfected, the start of the new levels has began, and the pause menu has been thrown out and the start of its remake is in with skills being the only menu that was started, as it now has the baseline function of allowing players to set their skills.